Checkpoint before assistant change: Implement calculations and data extraction for simulation output
Implements _getDistanceMeters and _parseMetadata in wasm.js, and migrates DataQuery and SummarizedResult to summarize.js.
