+++ editor/index.html
@@ -228,6 +228,7 @@ end simulation
           "model": "/js/model.js?v=EPOCH",
           "results": "/js/results.js?v=EPOCH",
           "run": "/js/run.js?v=EPOCH",
+          "summarize": "/js/summarize.js?v=EPOCH",
           "viz": "/js/viz.js?v=EPOCH",
           "wasm": "/js/wasm.js?v=EPOCH"
         }
